@SimonDanisch 2018-08-16
If we can turn CuDynamicSharedMem into CuStaticMem, this could easily live in GPUArrays ;)