@jekbradbury 2018-08-20
This should be ready. It now uses static shared memory, so it could potentially be moved to GPUArrays, but would probably need to be reworked first (like the original mapreduce kernel) to be compatible with Transpiler.