@MikeInnes 2018-08-26
Awesome work. Does one of those old version benchmarks really say 11 seconds? Either way, this is clearly a great improvement.

I imagine we are testing reduction already, but it'd be useful to test this kernel is getting called rather than the generic one. I'm not immediately sure how best to do that though.