@jekbradbury 2018-08-26
Yep, 11 seconds comes from a linear, serial scan over 33 million elements. I’m also unsure how to test that a particular kernel is being called—would be straightforward with Cassette but that’s probably not an acceptable test dependency.