@skariel 2018-09-16
I'm having the "too many resources" problem, just submitted a new issue here: #131